Remove sensitive data from config.yml.org file
The config.yml.org file was deleted to wipe out sensitive data, including uniquely identifiable information and API URLs for webhooks, that was publicly exposed. This step is crucial to maintain the confidentiality, integrity, and availability of the data, which could be exploited in malicious activities if left exposed. The config file will be updated with the safe data and new system environment variables will be used moving forward.