"Added search functionality and optimized GUI interactions"
Implemented a new search feature in the Shop GUI by adding a searchPattern in ShopGUI.java, allowing users to query for specific items. Modified the ShopCommand.java, StoredItemsGUI.java, SeeOfferedItemsGUI.java, and BuyGUI.java, to improve interaction with the GUI by passing the player directly to the ShopGUI constructor, enabling the user to see and interact with the search results immediately. Also, a new dependency (net.wesjd:anvilgui:1.9.0-SNAPSHOT) was added in build.gradle for supporting anvil GUI operations.